Make component for contributor
