+++ app/components/contributor-list/template.hbs
@@ -1,0 +1,1 @@
+{{contributorList}}
