+++ app/components/dashboard-item/component.js
@@ -1,11 +1,8 @@
 import { computed } from '@ember/object';
-import { inject as service } from '@ember/service';
 import moment from 'moment';
 import Component from '@ember/component';
-import contributorList from '../../utils/contributor-list';
 
 export default Component.extend({
-    i18n: service(),
     ancestry: computed('node', 'node.{parent,root}', function() {
         const rootId = this.get('node._internalModel.__data.links.relationships.root.errors.length') ? undefined : this.get('node.root.id');
         const parentId = this.get('node._internalModel.__data.links.relationships.parent.errors.length') ? undefined : this.get('node.parent.id');
@@ -30,8 +27,5 @@ export default Component.extend({
     }),
     date: computed('node.dateModified', function() {
         return moment(this.get('node.dateModified')).utc().format('YYYY-MM-DD h:mm A');
-    }),
-    contributors: computed('node.contributors', function() {
-        return contributorList(this.get('node'), this.get('i18n').t('general.and'));
     }),
 });
