+++ app/components/dashboard-item/template.hbs
@@ -17,7 +17,7 @@
                     </div>
                 </div>
                 <div class="col-sm-3 col-md-3 p-v-xs">
-                    <div class="quick-search-col">{{contributors}}</div>
+                    <div class="quick-search-col">{{contributor-list node=node}}</div>
                 </div>
                 <div class="col-sm-3 col-md-3 p-v-xs">
                     <div class="quick-search-col">{{date}}</div>
