+++ app/components/noteworthy-and-popular-project/component.js
@@ -1,15 +1,11 @@
 import { computed } from '@ember/object';
 import Component from '@ember/component';
-import contributorList from '../../utils/contributor-list';
 
 export default Component.extend({
     didInsertElement() {
         this._super(...arguments);
         this.$('.prevent-overflow').tooltip();
     },
-    contributors: computed('project', function() {
-        return contributorList(this.get('project'));
-    }),
     compactDescription: computed('project.description', function() {
         const desc = this.get('project.description');
         if (!desc) {
