--- tests/unit/utils/contributor-list-test.js
@@ -1,10 +1,0 @@
-import contributorList from 'ember-osf-web/utils/contributor-list';
-import { module, test } from 'qunit';
-
-module('Unit | Utility | contributor list');
-
-// Replace this with your real tests.
-test('it works', function(assert) {
-    const result = contributorList();
-    assert.ok(result);
-});
